update parse & formatter functions
